+++ src/senaite/core/schema/interfaces.py
@@ -74,6 +74,7 @@ class IRichTextField(IRichText):
     """Senaite rich text field
     """
 
+
 class IEmailField(ITextLine):
     """Input type "email" widget
     """
@@ -96,6 +97,11 @@ class IGPSCoordinatesField(IDict):
 
 class ISelectOtherField(INativeString):
     """Senaite SelectOther field
+    """
+
+
+class IMultiUploadField(IList):
+    """Senaite Multi Upload field
     """